move get_function_selector to mock
