+++ primitives/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Acala Developers"]
 edition = "2018"
 
 [dependencies]
-sha3 = { version = "0.9.1", default-features = false }
 bstringify = "0.1.2"
 serde = { version = "1.0.124", optional = true }
 codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
@@ -22,7 +21,6 @@ frame-support = { git = "https://github.com/paritytech/substrate", branch = "roc
 default = ["std"]
 std = [
 	"serde",
-	"sha3/std",
 	"codec/std",
 	"sp-runtime/std",
 	"sp-core/std",
