+++ runtime/common/src/precompile/dex.rs
@@ -257,7 +257,7 @@ where
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use primitives::get_function_selector;
+	use crate::precompile::mock::get_function_selector;
 
 	#[test]
 	fn function_selector_match() {
