+++ runtime/common/src/precompile/multicurrency.rs
@@ -157,7 +157,7 @@ fn vec_u8_from_str(b: &[u8]) -> Vec<u8> {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use primitives::get_function_selector;
+	use crate::precompile::mock::get_function_selector;
 
 	#[test]
 	fn function_selector_match() {
