+++ runtime/common/src/precompile/oracle.rs
@@ -86,7 +86,7 @@ fn vec_u8_from_price(value: Price) -> Vec<u8> {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use primitives::get_function_selector;
+	use crate::precompile::mock::get_function_selector;
 
 	#[test]
 	fn function_selector_match() {
