+++ runtime/common/src/precompile/schedule_call.rs
@@ -289,7 +289,7 @@ impl<
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use primitives::get_function_selector;
+	use crate::precompile::mock::get_function_selector;
 
 	#[test]
 	fn function_selector_match() {
